+++ crates/aiken-project/src/options.rs
@@ -1,9 +1,11 @@
 use aiken_lang::ast::Tracing;
+use std::path::PathBuf;
 
 pub struct Options {
     pub code_gen_mode: CodeGenMode,
     pub tracing: Tracing,
     pub env: Option<String>,
+    pub blueprint_path: PathBuf,
 }
 
 impl Default for Options {
@@ -12,6 +14,7 @@ impl Default for Options {
             code_gen_mode: CodeGenMode::NoOp,
             tracing: Tracing::silent(),
             env: None,
+            blueprint_path: PathBuf::from("plutus.json"),
         }
     }
 }