+++ crates/aiken/src/cmd/blueprint/policy.rs
@@ -7,6 +7,18 @@ pub struct Args {
     /// Path to project
     directory: Option<PathBuf>,
 
+    /// Optional path to the blueprint file to be used as input.
+    ///
+    /// [default: plutus.json]
+    #[clap(
+        short,
+        long = "in",
+        value_parser,
+        value_name = "FILEPATH",
+        verbatim_doc_comment
+    )]
+    input: Option<PathBuf>,
+
     /// Name of the validator's module within the project. Optional if there's only one validator
     #[clap(short, long)]
     module: Option<String>,
@@ -19,6 +31,7 @@ pub struct Args {
 pub fn exec(
     Args {
         directory,
+        input,
         module,
         validator,
     }: Args,
@@ -36,7 +49,7 @@ pub fn exec(
 
         let title = title.as_ref().or(validator.as_ref());
 
-        let policy = p.policy(title)?;
+        let policy = p.policy(title, p.blueprint_path(input.as_deref()).as_path())?;
 
         println!("{}", policy);